requirements.txt
transformers
evaluate
rouge_score
nltk
pycocoevalcap
scikit-learn
pyyaml